backend/api/src/requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
pymongo==3.11.4
python-dotenv==0.17.1
Werkzeug==2.0.1
